+++ src/main/scala/org/podval/tools/testing/framework/ScalaCheck.scala
@@ -9,7 +9,7 @@ object ScalaCheck extends FrameworkDescriptor(
   displayName = "ScalaCheck",
   group = "org.scalacheck",
   artifact = "scalacheck",
-  versionDefault = "1.18.0",
+  versionDefault = "1.18.1",
   className = "org.scalacheck.ScalaCheckFramework",
   sharedPackages = List("org.scalacheck")
 ):