+++ src/main/scala/org/podval/tools/testing/framework/Specs2.scala
@@ -11,7 +11,7 @@ object Specs2 extends FrameworkDescriptor(
   displayName = "Spec2",
   group = "org.specs2",
   artifact = "specs2-core",
-  versionDefault = "5.5.3",
+  versionDefault = "5.5.8",
   className = "org.specs2.runner.Specs2Framework",
   sharedPackages = List("org.specs2.runner") // TODO more?
 ):